small warning fix for alpha
2006-09-09  Dalibor Topic  <robilad@kaffe.org>

        * kaffe/kaffevm/classMethod.h (struct Hjava_lang_Class):
        Changed type of impl_index to unsigned int as its used as
        an array index.

        * kaffe/kaffevm/jni/jni-callmethod.c (getMethodFunc):
        Cast implementors to uintptr_t to fix warning on alpha-linux.

+++ kaffe/kaffevm/jni/jni-callmethod.c
@@ -36,7 +36,7 @@ getMethodFunc (Method* meth, Hjava_lang_Object *obj)
     implementors = meth->class->implementors;
     clazz = OBJECT_CLASS(obj);
 
-    assert (implementors != NULL && clazz->impl_index <= (int)implementors[0]);
+    assert (implementors != NULL && clazz->impl_index <= (uintptr_t) implementors[0]);
 
     return implementors[clazz->impl_index][meth->idx + 1]; 	
   } else {
